Fix group and item adapters
Make group and item adapters extend
the collection adapter so that a trailing
slash is appended to the end of the urls
that ember uses. Without the trailing slash,
the DRF API will throw an error.

Also, since top-level group and item routes are
in the API, remove custom buildURL methods from
group and item adapters.